pyproject.toml
[tool.poetry]
name = "langchain-chatmodel-prompt templates"
version = "0.1.0"
description = ""
authors = ["Nasir Abbas <akhtarabbas.islamicreaction@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-core = "^0.3.9"
langchain-google-genai = "^2.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
